walker setting

+++ parts/brain.py
@@ -106,7 +106,7 @@ class Brain:
 
         images = self.sensor.get()
         # image = np.clip(images[-1] - self.base_image, 0, 1500)
-        image = np.clip(images[-1], 0, 2500)
+        image = np.clip(images[-1], 0, 2500).astype(np.float64)
         image /= float(2500)
         x_from = self.right_arm_cx - int(self.right_arm_x_range / 2)
         x_end = self.right_arm_cx + int(self.right_arm_x_range / 2)
